feat(cz-git): reduce editor dev
