--- packages/@cz-git/editor/README.md
@@ -1,11 +1,0 @@
-# `@cz-git/editor`
-
-> TODO: description
-
-## Usage
-
-```
-const editor = require('@cz-git/editor');
-
-// TODO: DEMONSTRATE API
-```
