--- packages/@cz-git/editor/package.json
@@ -1,34 +1,0 @@
-{
-  "name": "@cz-git/editor",
-  "version": "1.1.1",
-  "description": "open editor for cz-git",
-  "keywords": [
-    "cz-git",
-    "editor"
-  ],
-  "homepage": "https://github.com/Zhengqbbb/cz-git/tree/main/packages/@cz-git/editor#readme",
-  "bugs": {
-    "url": "https://github.com/Zhengqbbb/cz-git/issues"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/Zhengqbbb/cz-git.git",
-    "directory": "packages/@cz-git/editor"
-  },
-  "license": "MIT",
-  "author": "Zhengqbbb <1074059947@qq.com>",
-  "main": "lib/index.js",
-  "directories": {
-    "lib": "lib",
-    "test": "__tests__"
-  },
-  "files": [
-    "lib"
-  ],
-  "scripts": {
-    "test": "echo \"Error: run tests from root\" && exit 1"
-  },
-  "publishConfig": {
-    "access": "public"
-  }
-}
