--- packages/@cz-git/editor/src/index.ts
@@ -1,1 +1,0 @@
-console.log(1);
