--- packages/@cz-git/editor/tsconfig.json
@@ -1,9 +1,0 @@
-{
-  "extends": "./tsconfig.base.json",
-  "compilerOptions": {
-    "rootDir": "./src",
-    "outDir": "./lib"
-  },
-  "include": ["./src"],
-  "exclude": ["./lib/**/*"]
-}
