--- packages/@cz-git/types/README.md
@@ -1,4 +1,0 @@
-# `@cz-git/types`
-
-> fork by [@commitlint/types](https://github.com/conventional-changelog/commitlint/tree/v16.2.2/@commitlint/types)
-> provide cz-git Userconfig
