--- packages/@cz-git/types/package.json
@@ -1,32 +1,0 @@
-{
-  "name": "@cz-git/types",
-  "version": "1.1.1",
-  "description": "Shared and fork by commitlint types to cz-git",
-  "keywords": [
-    "cz-git"
-  ],
-  "homepage": "https://github.com/Zhengqbbb/cz-git#readme",
-  "bugs": {
-    "url": "https://github.com/Zhengqbbb/cz-git/issues"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/Zhengqbbb/cz-git",
-    "directory": "packages/@cz-git/types"
-  },
-  "license": "MIT",
-  "author": "Zhengqbbb <zhengqbbb@gmail.com> (https://github.com/Zhengqbbb)",
-  "main": "lib/index.js",
-  "types": "lib/index.d.ts",
-  "files": [
-    "lib"
-  ],
-  "scripts": {
-    "build": "yarn clean && tsc -b",
-    "clean": "rimraf *.tsbuildinfo lib",
-    "dev": "tsc -b --watch"
-  },
-  "publishConfig": {
-    "access": "public"
-  }
-}
