--- packages/@cz-git/types/src/ensure.ts
@@ -1,13 +1,0 @@
-export type TargetCaseType =
-  | "camel-case"
-  | "kebab-case"
-  | "snake-case"
-  | "pascal-case"
-  | "start-case"
-  | "upper-case"
-  | "uppercase"
-  | "sentence-case"
-  | "sentencecase"
-  | "lower-case"
-  | "lowercase"
-  | "lowerCase";
